requirements.txt
torch-directml        
transformers          
gym                    
gym-minigrid          
mlagents              
stable-baselines3[extra]
tensorboard           
numpy
opencv-python 
pyglet 
pygame